Modularize chat into component.
