+++ frontend/src/routes/dashboard/+page.svelte
@@ -9,6 +9,7 @@
 	import { createAvailability, createTask } from '$lib/model';
 	import PollCreationModal from './PollCreationModal.svelte';
 	import { goto } from '$app/navigation';
+	import Chat from './Chat.svelte';
 
 	let start, end;
 	let availableTimes = [];
@@ -20,8 +21,7 @@
 	let taskMsg = writable('');
 	let taskInput = '';
 	let assignedInput = '';
-
-	let isPoll = writable(false);
+	let isPoll = false;
 
 	onMount(async () => {
 		if (!group) {
@@ -109,33 +109,9 @@
 			return currentTasks;
 		});
 	}
-	// for chat box
-	let messages = [];
-	let content = '';
-	/**
-	 * Send a message to the chat.
-	 */
-	function handleSendMessage() {
-		if (content.trim() !== '') {
-			messages = [...messages, { text: content, sender: 'user' }];
-			content = content.trim();
-			// sendMessage($groupId, content);
-			content = '';
-			// need to handle Fetch messages
-		}
-	}
-	/**
-	 * send message if hit enter
-	 * @param event
-	 */
-	function handleKeyPress(event) {
-		if (event.key === 'Enter') {
-			handleSendMessage();
-		}
-	}
 
 	function openPoll() {
-		isPoll.set(true);
+		isPoll = true;
 	}
 
 	async function saveAllAvailabilities() {
@@ -201,33 +177,7 @@
 			</button>
 		</div>
 
-		<div class="chatbox">
-			<div class="messages">
-				{#each messages as message (message.text)}
-					<div class:message class:message.sender={message.sender}>
-						{#if message.sender === 'user'}
-							<strong class="user-message">You:</strong> {message.text}
-						{:else if message.sender === 'system'}
-							<em class="system-message">{message.text}</em>
-						{/if}
-					</div>
-				{/each}
-			</div>
-			<div class="poll">
-				{#if $isPoll}
-					<PollCreationModal />
-				{/if}
-			</div>
-			<div class="input-bar">
-				<input
-					class="input"
-					bind:value={content}
-					placeholder="Type your message..."
-					on:keydown={handleKeyPress}
-				/>
-				<button on:click={handleSendMessage} on:keyup={handleSendMessage}>Send Message</button>
-			</div>
-		</div>
+		<Chat {group} bind:isPoll={isPoll}/>
 
 		<div class="calendar-container">
 			<span class="calendar-title">AVAILABILITY CALENDAR</span>
@@ -419,7 +369,7 @@
 		border: 2px solid transparent;
 	}
 
-	button {
+	:global(button) {
 		padding: 0.5rem 1rem;
 		background-color: #2774d0;
 		color: white;
@@ -430,7 +380,7 @@
 		cursor: pointer;
 	}
 
-	button:hover {
+	:global(button:hover) {
 		background-color: gray;
 		color: white;
 	}
@@ -464,54 +414,12 @@
 		font-weight: bold;
 	}
 
-	button[type='submit']:disabled {
+	:global(button[type='submit']:disabled) {
 		background-color: #ccc;
 		cursor: not-allowed;
 	}
-	/* chatbox style */
-	.chatbox {
-		display: flex;
-		flex-direction: column;
-		border: 2px solid #ccc;
-		padding: 4rem 6rem 2rem 6rem;
-		max-width: calc(90% - 10px);
-		height: 700px;
-		border-radius: 8px;
-		box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-		overflow-y: auto; /* Add scrollbar when content exceeds the height */
-		margin-right: 2rem;
-	}
 
-	.messages {
-		flex-grow: 1;
-		display: flex;
-		flex-direction: column-reverse; /* Reverse the order of messages */
-	}
-
-	.message {
-		margin: 8px 0;
-		padding: 8px;
-		background-color: #f0f0f0;
-		border-radius: 4px;
-	}
-
-	.user-message {
-		background-color: #e6f7ff;
-		text-align: right;
-	}
-
-	.system-message {
-		color: #888;
-		font-style: italic;
-	}
-
-	.input-bar {
-		display: flex;
-		align-items: center;
-		justify-content: space-between;
-		margin-top: 10px;
-	}
-	input {
+	:global(input) {
 		padding: 0.5rem;
 		margin-bottom: 0.5rem;
 		width: 80%;
@@ -523,7 +431,7 @@
 		border: 2px solid transparent;
 	}
 
-	button {
+	:global(button) {
 		flex-shrink: 0;
 	}
 </style>
